+++ vu/vmudm.h
@@ -10,7 +10,7 @@ INLINE void do_mudm(short* VD, short* VS, short* VT)
     for (i = 0; i < N; i++)
         ACC_H(i) = VS[i] >> 15;
     for (i = 0; i < N; i++)
-        ACC_M(i) = (acc[i] >> 16);
+        ACC_M(i) = (unsigned int)(acc[i]) >> 16;
     for (i = 0; i < N; i++)
         ACC_L(i) = acc[i];
     for (i = 0; i < N; i++)
